sudokuTS-22 Moved code not strictly belonging to DLX to its own class
Created a file `DLXHelpers`
